requirements.txt
requests==2.31.0
pandas==1.4.0
datetime==5.2
python-dotenv==0.21.0